TemporalAssistant

+++ src/jvmMain/kotlin/ru/capjack/tool/utils/assistant/TemporalAssistant.kt
@@ -5,9 +5,10 @@ import java.util.concurrent.TimeUnit
 
 actual interface TemporalAssistant : Assistant {
 	actual fun schedule(delayMillis: Int, code: () -> Unit): Cancelable
+	
 	actual fun repeat(delayMillis: Int, code: () -> Unit): Cancelable
 	
-	fun schedule(delay: Int, unit: TimeUnit, code: () -> Unit): Cancelable
+	fun schedule(delay: Long, unit: TimeUnit, code: () -> Unit): Cancelable
 	
-	fun repeat(delay: Int, unit: TimeUnit, code: () -> Unit): Cancelable
+	fun repeat(delay: Long, unit: TimeUnit, code: () -> Unit): Cancelable
 }
